pyproject.toml
[project]
name = "ai-openai-agent-prodready"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backoff>=2.2.1",
    "httpx>=0.28.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "structlog>=25.4.0",
]
